+++ sql/install/02-inspection.sql
@@ -14,33 +14,41 @@
  * limitations under the License.
  */
 
--- Index inspection functions (inspection.c)
+-- ---------- ---------- ---------- ---------- ---------- ---------- ----------
+-- Index Inspection Functions (inspection.c)
+-- ---------- ---------- ---------- ---------- ---------- ---------- ----------
 
+-- Availability: 1.0.0
 CREATE OR REPLACE FUNCTION h3_get_resolution(h3index) RETURNS integer
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_get_resolution(h3index) IS
 'Returns the resolution of the index';
 
+-- Availability: 1.0.0
 CREATE OR REPLACE FUNCTION h3_get_base_cell(h3index) RETURNS integer
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_get_base_cell(h3index) IS
 'Returns the base cell number of the index';
 
+-- Availability: 1.0.0
 CREATE OR REPLACE FUNCTION h3_is_valid(h3index) RETURNS bool
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_is_valid(h3index) IS
 'Returns true if the given H3Index is valid';
 
+-- Availability: 1.0.0
 CREATE OR REPLACE FUNCTION h3_is_res_class_iii(h3index) RETURNS bool
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_is_res_class_iii(h3index) IS
 'Returns true if this index has a resolution with Class III orientation';  
   
+-- Availability: 1.0.0
 CREATE OR REPLACE FUNCTION h3_is_pentagon(h3index) RETURNS bool
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_is_pentagon(h3index) IS
 'Returns true if this index represents a pentagonal cell';
 
+-- Availability: 3.5.0
 CREATE OR REPLACE FUNCTION h3_get_faces(h3index) RETURNS integer[]
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_get_faces(h3index) IS