+++ sql/install/05-regions.sql
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Bytes & Brains
+ * Copyright 2018-2019 Bytes & Brains
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,13 +14,17 @@
  * limitations under the License.
  */
 
--- Region functions (regions.c)
+-- ---------- ---------- ---------- ---------- ---------- ---------- ----------
+-- Region Functions (regions.c)
+-- ---------- ---------- ---------- ---------- ---------- ---------- ----------
 
+-- Availability: 0.2.0
 CREATE OR REPLACE FUNCTION h3_polyfill(exterior polygon, holes polygon[], resolution integer DEFAULT 1) RETURNS SETOF h3index
-    AS 'h3' LANGUAGE C IMMUTABLE PARALLEL SAFE; -- NOT STRICT
+    AS 'h3' LANGUAGE C IMMUTABLE CALLED ON NULL INPUT PARALLEL SAFE; -- NOT STRICT
     COMMENT ON FUNCTION h3_polyfill(exterior polygon, holes polygon[], resolution integer) IS
 'Takes an exterior polygon [and a set of hole polygon] and returns the set of hexagons that best fit the structure';
 
+-- Availability: 3.5.0
 CREATE OR REPLACE FUNCTION h3_set_to_multi_polygon(h3index[], OUT exterior polygon, OUT holes polygon[]) RETURNS SETOF record
     AS 'h3' LANGUAGE C IMMUTABLE STRICT PARALLEL SAFE;
     COMMENT ON FUNCTION h3_set_to_multi_polygon(h3index[]) IS