+++ src/opclass_btree.c
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2018-2019 Bytes & Brains
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *	   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include <postgres.h>		 // Datum, etc.
+#include <fmgr.h>			 // PG_FUNCTION_ARGS, etc.
+
+#include <h3api.h> // Main H3 include
+#include "extension.h"
+
+PG_FUNCTION_INFO_V1(h3index_cmp);
+
+Datum
+h3index_cmp(PG_FUNCTION_ARGS)
+{
+	H3Index    *a = PG_GETARG_H3_INDEX_P(0);
+	H3Index    *b = PG_GETARG_H3_INDEX_P(1);
+
+	if (*a > *b)
+		PG_RETURN_INT32(1);
+	else if (*a == *b)
+		PG_RETURN_INT32(0);
+	else
+		PG_RETURN_INT32(-1);
+}