+++ src/type.c
@@ -22,19 +22,47 @@
 #include <h3api.h> // Main H3 include
 #include "extension.h"
 
+// conversion
 PG_FUNCTION_INFO_V1(h3index_in);
 PG_FUNCTION_INFO_V1(h3index_out);
 PG_FUNCTION_INFO_V1(h3index_to_bigint);
 PG_FUNCTION_INFO_V1(bigint_to_h3index);
+
+// b-tree
 PG_FUNCTION_INFO_V1(h3index_eq);
 PG_FUNCTION_INFO_V1(h3index_ne);
 PG_FUNCTION_INFO_V1(h3index_lt);
 PG_FUNCTION_INFO_V1(h3index_le);
 PG_FUNCTION_INFO_V1(h3index_gt);
 PG_FUNCTION_INFO_V1(h3index_ge);
-PG_FUNCTION_INFO_V1(h3index_cmp);
-PG_FUNCTION_INFO_V1(h3index_hash);
 
+// r-tree
+PG_FUNCTION_INFO_V1(h3index_overlaps);
+PG_FUNCTION_INFO_V1(h3index_contains);
+PG_FUNCTION_INFO_V1(h3index_contained_by);
+
+// static helpers
+static int
+containment(H3Index * a, H3Index * b)
+{
+	int			aRes = h3GetResolution(*a);
+	int			bRes = h3GetResolution(*b);
+	H3Index		aParent = h3ToParent(*a, bRes);
+	H3Index		bParent = h3ToParent(*b, aRes);
+
+	/* a contains b */
+	if (*a == bParent)
+		return 1;
+
+	/* a contained by b */
+	if (*b == aParent)
+		return -1;
+
+	/* no overlap */
+	return 0;
+}
+
+// textual input/output functions
 Datum
 h3index_in(PG_FUNCTION_ARGS)
 {
@@ -57,6 +85,7 @@ h3index_out(PG_FUNCTION_ARGS)
 	PG_RETURN_CSTRING(str);
 }
 
+// bigint conversion functions
 Datum
 h3index_to_bigint(PG_FUNCTION_ARGS)
 {
@@ -75,6 +104,7 @@ bigint_to_h3index(PG_FUNCTION_ARGS)
 	PG_RETURN_H3_INDEX_P(h3index);
 }
 
+// b-tree operators
 Datum
 h3index_eq(PG_FUNCTION_ARGS)
 {
@@ -129,25 +159,30 @@ h3index_ge(PG_FUNCTION_ARGS)
 	PG_RETURN_BOOL(*a >= *b);
 }
 
+// r-tree operators
 Datum
-h3index_cmp(PG_FUNCTION_ARGS)
+h3index_overlaps(PG_FUNCTION_ARGS)
 {
 	H3Index    *a = PG_GETARG_H3_INDEX_P(0);
 	H3Index    *b = PG_GETARG_H3_INDEX_P(1);
 
-	if (*a > *b)
-		PG_RETURN_INT32(1);
-	else if (*a == *b)
-		PG_RETURN_INT32(0);
-	else
-		PG_RETURN_INT32(-1);
+	PG_RETURN_BOOL(containment(a, b) != 0);
 }
 
 Datum
-h3index_hash(PG_FUNCTION_ARGS)
+h3index_contains(PG_FUNCTION_ARGS)
 {
-	H3Index    *index = PG_GETARG_H3_INDEX_P(0);
-	uint32		hash = hash_any((unsigned char *) index, sizeof(H3Index));
+	H3Index    *a = PG_GETARG_H3_INDEX_P(0);
+	H3Index    *b = PG_GETARG_H3_INDEX_P(1);
 
-	PG_RETURN_INT32(hash);
+	PG_RETURN_BOOL(containment(a, b) > 0);
+}
+
+Datum
+h3index_contained_by(PG_FUNCTION_ARGS)
+{
+	H3Index    *a = PG_GETARG_H3_INDEX_P(0);
+	H3Index    *b = PG_GETARG_H3_INDEX_P(1);
+
+	PG_RETURN_BOOL(containment(a, b) < 0);
 }