+++ test/sql/extension.sql
@@ -4,4 +4,5 @@
 -- TEST h3_get_extension_version
 --
 
-SELECT h3_get_extension_version() ~ '^[0-9]+\.[0-9]+\.[0-9]+$';
+SELECT h3_get_extension_version() ~ '^[0-9]+\.[0-9]+\.[0-9]+$'
+    OR h3_get_extension_version() = 'unreleased';