+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Because H3-pg is versioned in lockstep with the H3 core library, please
 avoid adding features or APIs which do not map onto the
 [H3 core API](https://uber.github.io/h3/#/documentation/api-reference/).
 
+## [Unreleased]
+### Added
+- Add operators `&&`, `@>` and `<@` for overlap, contains and contained by respectively.
+
 ## [3.6.0] - 2019-10-07
 ### Added
 - `h3_to_center_child` function to find center child at given resolution.